More cleanup

+++ src/main/kotlin/world/cepi/itemextension/item/traits/list/ArmorTrait.kt
@@ -25,7 +25,7 @@ data class ArmorTrait(
     }
 
     companion object {
-        fun applyToDamage(armor: Double, damage: Double): Double =
+        fun applyToDamage(armor: Double, damage: Float): Double =
             (damage - armor).coerceAtLeast(0.0)
     }
